+++ submission.js
@@ -56,7 +56,9 @@ function runExactRouteTestCases(pGraphObj, resArr)
 function runShortestRouteTestCases(pGraphObj, resArr)
 {
 	var case8 = shortestRoute.findRoute(pGraphObj, "A", "C");
-	//var case9 = shortestRoute.findRoute(pGraphObj, "B", "B");
+	console.log("");
+	console.log(case8);
 	
+	//var case9 = shortestRoute.findRoute(pGraphObj, "B", "B");
 	//resArr.push(case8, case9);
 }