+++ changelog.md
@@ -1,9 +1,24 @@
 # Changelog
 
-**./src/shortest-route.js - loopPathfinding**
-* Removed 'iterationsMade' variable.
-* Removed `if (iterationsMade === 5)` structure.
-* Loop will abort when all nodes have been visited.
-* After loop finishes:
-	* Reset node sort order.
-	* Display nodes table.
+**./src/common/dsktra-tasks.js**
+* Removed commented out 'distanceFromStart' assignment from 'setNodeTable'
+* Removed 'closedRoute' property from 'defineDijkstraInfo'
+* Removed the 'setClosedRoute' function.
+
+---
+
+**./src/shortest-route.js**
+* Renamed 'loopPathfinding' function to 'visitOtherNodes'
+* Wrote new function 'checkRouteSuccessful'
+	* Used to check if pathfinding was successful.
+	* Possible route between start and end nodes.
+	* If successful, output distance.
+	* Otherwise, output message.
+* Changes to 'findShortestRoute'
+	* Removed 'dsktraTasks.setNodes' call.
+	* Declared 'shortRouteRes' variable.
+	* If the start or end nodes are missing, output message.
+	* Return 'shortRouteRes'
+
+**./submission.js - runShortestRouteTestCases**
+* Output 'case8' to console.